all: Run 'cargo clippy --fix'
This applies all clippy lints that are marked as MachineApplicable, i.e.,
safe to apply without changing code semantics

+++ graph/src/data/store/scalar/bigint.rs
@@ -73,7 +73,7 @@ mod big_int {
         }
 
         pub fn bits(&self) -> usize {
-            self.0.bits() as usize
+            self.0.bits()
         }
 
         pub(in super::super) fn inner(self) -> num_bigint::BigInt {
